# internal/handler/docs/appapi_swagger.yaml
basePath: /api/v1
definitions:
...
  role.RoleInfo:
    properties:
      code:
        description: role code, must be alpha numeric
        example: ADMIN
        type: string
      id:
        description: role id
        example: 1
        type: integer
      name:
        description: role name
        example: admin
        type: string
    type: object
...
  types.Response:
    properties:
      code:
        example: 2000
        type: integer
      data: {}
      err:
        type: string
      msg:
        example: operation success
        type: string
    type: object
  user.CreateUserOption:
    properties:
      email:
        description: new email
        example: jack@google.com
        type: string
      password:
        description: new password
        example: "123456"
        type: string
      roles:
        description: new roles
        items:
          type: string
        type: array
      username:
        description: new username
        example: jack
        type: string
    type: object
  user.Info:
    properties:
      createdAt:
        example: 947416200
        type: integer
      email:
        example: jacklove@lol.com
        type: string
      roles:
        items:
          $ref: '#/definitions/role.RoleInfo'
        type: array
      username:
        example: jack
        type: string
      uuid:
        example: 55BBA4ED-18D3-790F-EABF-A5330E527586
        type: string
    type: object
  user.SaveUserDetailOption:
    properties:
      email:
        description: new email
        example: jack@google.com
        type: string
      password:
        description: new password
        example: "123456"
        type: string
      roles:
        description: new roles
        items:
          type: string
        type: array
      username:
        description: new username
        example: jack
        type: string
      uuid:
        example: 55BBA4ED-18D3-790F-EABF-A5330E527586
        type: string
    type: object
  user.UpdateInfoOption:
    properties:
      email:
        description: new email
        example: jack@google.com
        type: string
      password:
        description: new password
        example: "123456"
        type: string
      username:
        description: new username
        example: jack
        type: string
    type: object
info:
  contact:
    name: dstgo
    url: https://github.com/dstgo
  description: |-
    app http api documentation, use the Bearer Token to authenticate
    It should be noted that when using swagger doc for API debugging, the Token needs to be manually prefixed with Bearer.
  license:
    name: MIT LICENSE
    url: https://mit-license.org/
  title: App Internal API Documentation
  version: v1.0.0
paths:
...
  /user/admin/create:
    post:
      consumes:
      - application/json
      description: |-
        [admin]
        create new user
      parameters:
      - description: CreateUserOption
        in: body
        name: createOpt
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserOption'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: CreateUser
      tags:
      - user
  /user/admin/list:
    get:
      consumes:
      - application/json
      description: |-
        [admin]
        get specific user list
      parameters:
      - description: if is reverse order
        example: true
        in: query
        name: desc
        type: boolean
      - description: specified field
        example: email
        in: query
        name: order
        type: string
      - description: page
        example: 1
        in: query
        name: page
        type: integer
      - description: search text, should be of one username or email
        example: jacklove
        in: query
        name: search
        type: string
      - description: size of page
        example: 10
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/user.Info'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: GetUserInfoList
      tags:
      - user
  /user/admin/profile:
    get:
      consumes:
      - application/json
      description: |-
        [admin]
        get specified user information
      parameters:
      - in: query
        name: uuid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: GetSpecUserInfo
      tags:
      - user
    post:
      consumes:
      - application/json
      description: |-
        [admin]
        save specified user information
      parameters:
      - description: updateInfoOption
        in: body
        name: updateInfoOption
        required: true
        schema:
          $ref: '#/definitions/user.SaveUserDetailOption'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: SaveUser
      tags:
      - user
  /user/admin/remove:
    delete:
      consumes:
      - application/json
      description: |-
        [admin]
        Remove the specific user
      parameters:
      - in: query
        name: uuid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: RemoveUser
      tags:
      - user
  /user/profile:
    get:
      consumes:
      - application/json
      description: |-
        [user]
        get own user info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/user.Info'
              type: object
      security:
      - BearerAuth: []
      summary: GetUserInfo
      tags:
      - user
    post:
      consumes:
      - application/json
      description: |-
        [user]
        update own user info
      parameters:
      - description: updateInfoOption
        in: body
        name: updateInfoOption
        required: true
        schema:
          $ref: '#/definitions/user.UpdateInfoOption'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: UpdateUserInfo
      tags:
      - user
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"